Move sp files to `keyboard`/`settings`/`storage` subfolders
