+++ configure.py
@@ -300,34 +300,40 @@ code_in_files = {
         os.path.join('payload', 'revolution', 'os', 'OSMemory.S'),
         os.path.join('payload', 'revolution', 'os', 'OSThread.S'),
         os.path.join('payload', 'revolution', 'os', 'OSThread.c'),
-        os.path.join('payload', 'sp', 'BaseSettings.c'),
-        os.path.join('payload', 'sp', 'ClientSettings.c'),
         os.path.join('payload', 'sp', 'Fatal.c'),
-        os.path.join('payload', 'sp', 'FatStorage.c'),
         os.path.join('payload', 'sp', 'FormattingCodes.c'),
         os.path.join('payload', 'sp', 'FlameGraph.c'),
         os.path.join('payload', 'sp', 'Host.c'),
-        os.path.join('payload', 'sp', 'IniReader.c'),
         os.path.join('payload', 'sp', 'IOSDolphin.c'),
-        os.path.join('payload', 'sp', 'IOSKeyboard.c'),
-        os.path.join('payload', 'sp', 'Keyboard.c'),
-        os.path.join('payload', 'sp', 'LogFile.c'),
+        # Keyboard module
+        os.path.join('payload', 'sp', 'keyboard', 'Keyboard.c'),
+        os.path.join('payload', 'sp', 'keyboard', 'SIKeyboard.c'),
+        os.path.join('payload', 'sp', 'keyboard', 'IOSKeyboard.c'),
+        #
         os.path.join('payload', 'sp', 'Memory.c'),
         os.path.join('payload', 'sp', 'Net.c'),
-        os.path.join('payload', 'sp', 'NetStorage.c'),
-        os.path.join('payload', 'sp', 'NetStorageClient.c'),
         os.path.join('payload', 'sp', 'Patcher.c'),
         os.path.join('payload', 'sp', 'Payload.c'),
-        os.path.join('payload', 'sp', 'Sdi.c'),
-        os.path.join('payload', 'sp', 'SIKeyboard.c'),
+        # Settings module
+        os.path.join('payload', 'sp', 'settings', 'BaseSettings.c'),
+        os.path.join('payload', 'sp', 'settings', 'ClientSettings.c'),
+        os.path.join('payload', 'sp', 'settings', 'IniReader.c'),
+        #
         os.path.join('payload', 'sp', 'Slab.c'),
         os.path.join('payload', 'sp', 'Stack.S'),
         os.path.join('payload', 'sp', 'Stack.c'),
         os.path.join('payload', 'sp', 'StackTrace.c'),
-        os.path.join('payload', 'sp', 'Storage.c'),
+        # Storage module
+        os.path.join('payload', 'sp', 'storage', 'FatStorage.c'),
+        os.path.join('payload', 'sp', 'storage', 'LogFile.c'),
+        os.path.join('payload', 'sp', 'storage', 'NetStorage.c'),
+        os.path.join('payload', 'sp', 'storage', 'NetStorageClient.c'),
+        os.path.join('payload', 'sp', 'storage', 'Sdi.c'),
+        os.path.join('payload', 'sp', 'storage', 'Storage.c'),
+        os.path.join('payload', 'sp', 'storage', 'Usb.c'),
+        os.path.join('payload', 'sp', 'storage', 'UsbStorage.c'),
+        #
         os.path.join('payload', 'sp', 'Tcp.c'),
-        os.path.join('payload', 'sp', 'Usb.c'),
-        os.path.join('payload', 'sp', 'UsbStorage.c'),
         os.path.join('payload', 'sp', 'Yaz.c'),
         os.path.join('vendor', 'arith64.c'),
         os.path.join('vendor', 'ff', 'diskio.c'),
